Fix bugs at Sources in Student Project, Adjust and improve the Review Page for Instructor

+++ BE/src/main/java/com/evidencepilot/dto/response/GraphResponse.java
@@ -53,6 +53,6 @@ public record GraphResponse(
 
     public record ClaimStatsResponse(
         int totalClaims,
-        Map<String, Long> byFunctionalType
+        Map<String, Double> byFunctionalType
     ) {}
 }
